Add automated test.
Make application testable by adding success message element to DOM that is hidden when page loads but becomes visible when XHR upload to bucket finishes. Add automated test that uses Puppeteer to select a test file (stored in this repo under test/1x1.png) and click the upload button to upload it. The test is considered successful as long as the XHR for the file upload succeeds. Added GitHub Actions workflow that occasionally clears out the contents of the bucket which was created only for this automated test.