Add automated test (#335)
* Add automated test.

Make application testable by adding success message element to DOM that is hidden when page loads but becomes visible when XHR upload to bucket finishes. Add automated test that uses Puppeteer to select a test file (stored in this repo under test/1x1.png) and click the upload button to upload it. The test is considered successful as long as the XHR for the file upload succeeds. Added GitHub Actions workflow that occasionally clears out the contents of the bucket which was created only for this automated test.

* Add step to 'run tests' GHA workflow to install Puppeteer libs.

* npm ci instead of npm install

* both front end and back end npm ci

* Longer test timeouts.

* Switch from default Puppeteer launch to launch with headless set to 'new'.

* Longer timeouts.

* Longer timeouts.

* Add step in GHA workflow to build front end before running tests.

* Add env vars to GHA workflow using secrets.

* Fix test GHA workflow to fail when it should.

* Fix test logs. Fix test GHA workflow env var.